+++ src/api/render/rce_internal.h
@@ -23,7 +23,7 @@ GPU_HIDE
 void
 gpuSetRenderVertexBuffer(GPURenderPassEncoder *pass,
                          GPUBuffer            *buf,
-                         size_t                off,
+                         uint64_t              off,
                          uint32_t              index);
 
 GPU_HIDE
@@ -42,7 +42,7 @@ GPU_HIDE
 void
 gpuSetRenderFragmentBuffer(GPURenderPassEncoder *pass,
                            GPUBuffer            *buf,
-                           size_t                off,
+                           uint64_t              off,
                            uint32_t              index);
 
 GPU_HIDE
